Update file structure and fix css
