+++ client/src/Scheduling/SchedulingContainer.jsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import '../css/Scheduling/SchedulingContainer.css'
-import Toolbar from './Toolbar';
-import ScheduleInterface from './ScheduleView';
+import Toolbar from './Toolbar/Toolbar';
+import ScheduleView from './ScheduleView/ScheduleView';
 
 function SchedulingContainer() {
   return (
-    <>
+    <div className='scheduling-container'>
       { <Toolbar /> }
-      { <ScheduleInterface /> }
-    </>
+      { <ScheduleView /> }
+    </div>
   )
 }
