+++ client/src/css/Scheduling/ScheduleView/ScheduleView.css
@@ -2,7 +2,7 @@
     width: 75%;
     height: 80%;
     margin-top: 10.7%;
-    margin-left: 23%;
+    margin-left: 0;
     display: flex;
     flex-direction: column;
 }
