+++ client/src/css/Scheduling/SchedulingContainer.css
@@ -1,0 +1,5 @@
+.scheduling-container {
+  display: flex;
+  gap: 7%;
+
+}
