+++ client/src/css/Scheduling/Toolbar/Toolbar.css
@@ -1,0 +1,7 @@
+.toolbar{
+    width: 25%;
+    height: 100%;
+    font-size: medium !important;
+}
+
+
